% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_functions.R
\name{add_ntile.cqtc}
\alias{add_ntile.cqtc}
\title{Add ntiles (quantiles) for a specific column across all subjects}
\usage{
\method{add_ntile}{cqtc}(obj, input_col = "CONC", n = 10, ntile_name = NULL)
}
\arguments{
\item{obj}{A cqtc object.}

\item{input_col}{The column to calculate quantiles over.}

\item{n}{The number of quantiles.}

\item{ntile_name}{The name of the quantile column.}
}
\value{
A cqtc object with the ntile_name colunn added.
}
\description{
Add ntiles (quantiles) for a specific column across all subjects
}
\examples{
head(add_ntile(dofetilide_cqtc, "CONC", 10))
}
