% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_class.R
\name{add_ntile}
\alias{add_ntile}
\title{Add ntiles (quantiles) for a specific column across all subjects}
\usage{
add_ntile(obj, input_col, n, ntile_name = NULL)
}
\arguments{
\item{obj}{A cqtc object.}

\item{input_col}{The column to calculate quantiles over.}

\item{n}{The number of quantiles.}

\item{ntile_name}{The name of the quantile column.}
}
\value{
A cqtc object with the ntile_name colunn added.
}
\description{
Add ntiles (quantiles) for a specific column across all subjects
}
