% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_make.R
\name{auto_cqtc}
\alias{auto_cqtc}
\title{Title}
\usage{
auto_cqtc(
  sdtm,
  treatment = NULL,
  pc_analyte = NULL,
  eg_analyte = c(QTCF = "QTCF", HR = "HR", RR = "RR", QT = "QTGDUR"),
  egcat = "CENTRAL ECG",
  ntime_method = "TPT",
  NTIME_lookup = NULL,
  silent = NULL,
  verbose = TRUE
)
}
\arguments{
\item{sdtm}{A sdtm object.}

\item{treatment}{The EXTRT.}

\item{pc_analyte}{The PCTESTCD.}

\item{eg_analyte}{The EGTESTCD.}

\item{egcat}{The EGCAT to filter observations for.}

\item{silent}{Suppress messages.}

\item{verbose}{Verbose messages.}
}
\value{
A cqtc object.
}
\description{
Title
}
