% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_class.R
\name{average_triplicates}
\alias{average_triplicates}
\title{Average triplicate observations}
\usage{
average_triplicates(obj)
}
\arguments{
\item{obj}{A cqtc object.}
}
\value{
A cqtc object with the HR, RR, QT, QTCF, DQTCF fields
averaged over ID, NTIME and CONC.
}
\description{
Average triplicate observations
}
\examples{
average_triplicates(quinidine_cqtc)
}
