% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_functions.R
\name{cqtc_add_baseline}
\alias{cqtc_add_baseline}
\title{Add baseline field for parameter across treatment groups}
\usage{
cqtc_add_baseline(
  obj,
  param = "QTCF",
  baseline_filter = "NTIME < 0",
  silent = NULL
)
}
\arguments{
\item{obj}{A cqtc object.}

\item{param}{The parameter to find the baseline for.}

\item{baseline_filter}{A filter term to identify unique baseline values.}

\item{silent}{Suppress messages.}
}
\value{
A cqtc object with the baseline columns added.
}
\description{
Add baseline field for parameter across treatment groups
}
