% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_exploration.R
\name{cqtc_decile_plot}
\alias{cqtc_decile_plot}
\title{Exploratory decile plot.}
\usage{
cqtc_decile_plot(
  obj,
  param = "DQTCF",
  n = 10,
  x_label = "concentration (ng/ml)",
  y_label = NULL,
  title = "",
  ...
)
}
\arguments{
\item{obj}{A cqtc object.}

\item{param}{The parameter to plot.}

\item{n}{The number of quantiles, defaults to 10.}

\item{x_label}{The x axis label.}

\item{y_label}{The y axis label.}

\item{title}{The plot title.}

\item{...}{Further parameters to geom_point.}
}
\value{
A ggplot object.
}
\description{
Exploratory decile plot.
}
\examples{
cqtc_decile_plot(dofetilide_cqtc)
}
