% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_gof.R
\name{cqtc_gof_plot}
\alias{cqtc_gof_plot}
\title{Goodness of fit plots for linear mixed effects model}
\usage{
cqtc_gof_plot(mod)
}
\arguments{
\item{mod}{Model result.}
}
\value{
A list of ggplot objects.
}
\description{
Goodness of fit plots for linear mixed effects model
}
