% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_exploration.R
\name{cqtc_model_fixed_effects}
\alias{cqtc_model_fixed_effects}
\title{Fixed effects of model}
\usage{
cqtc_model_fixed_effects(mod, method = "profile", level = 0.9)
}
\arguments{
\item{mod}{The linear mixed-effects model}

\item{level}{The confidence interval level.}
}
\value{
A data frame.
}
\description{
CI are likelihood profile CI
}
