% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_exploration.R
\name{cqtc_model_plot}
\alias{cqtc_model_plot}
\title{Plot predicted dQTcF over concentration}
\usage{
cqtc_model_plot(
  obj,
  mod,
  title = NULL,
  level = 0.9,
  size = 2,
  alpha = 0.1,
  lwd = 0.6,
  loess = FALSE
)
}
\arguments{
\item{obj}{A cqtc object.}

\item{mod}{A linear model.}

\item{title}{The plot title.}

\item{level}{The prediction interval.}

\item{size}{The point size.}

\item{alpha}{The point alpha value.}

\item{lwd}{The line width.}

\item{loess}{Show LOESS fit.}
}
\value{
A ggplot2 object.
}
\description{
Plot predicted dQTcF over concentration
}
