% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_exploration.R
\name{cqtc_model_table}
\alias{cqtc_model_table}
\title{Tabulate model parameters}
\usage{
cqtc_model_table(mod, method = "profile", level = 0.9, round = 3)
}
\arguments{
\item{mod}{The linear mixed-effects model}

\item{method}{The CI calculation method ("profile" or "boot")}

\item{level}{The confidence interval level.}

\item{round}{Number of decimal places in output.}
}
\value{
A data frame.
}
\description{
Tabulate model parameters
}
