% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_exploration.R
\name{cqtc_ntile_plot}
\alias{cqtc_ntile_plot}
\title{Exploratory decile plot.}
\usage{
cqtc_ntile_plot(
  obj,
  param = "DQTCF",
  n = 10,
  x_label = "concentration (ng/ml)",
  y_label = NULL,
  title = "",
  size = 2,
  alpha = 0.1,
  lwd = 0.6,
  loess = FALSE,
  lm = FALSE,
  ...
)
}
\arguments{
\item{obj}{A cqtc object.}

\item{param}{The parameter to plot.}

\item{n}{The number of quantiles, defaults to 10.}

\item{x_label}{The x axis label.}

\item{y_label}{The y axis label.}

\item{title}{The plot title.}

\item{size}{Point size.}

\item{alpha}{Alpha for points.}

\item{lwd}{Line width for point range.}

\item{loess}{Show LOESS, as logical.}

\item{lm}{Show linear regression, as logical.}

\item{...}{Further parameters to geom_point.}
}
\value{
A ggplot object.
}
\description{
Exploratory decile plot.
}
\examples{
cqtc_decile_plot(dofetilide_cqtc)
}
