% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_exploration.R
\name{cqtc_plot}
\alias{cqtc_plot}
\title{Exploratory c-QTc plot}
\usage{
cqtc_plot(
  obj,
  param = "QTCF",
  color = NULL,
  fit = TRUE,
  method = "loess",
  x_label = "concentration (ng/ml)",
  y_label = NULL,
  title = "",
  model = NULL,
  ...
)
}
\arguments{
\item{obj}{A cqtc object.}

\item{param}{The parameter to plot on the y axis.}

\item{color}{The column to be used for coloring.}

\item{fit}{Show regression fit.}

\item{method}{The method for geom_smooth.}

\item{x_label}{The x axis label.}

\item{y_label}{The y axis label.}

\item{title}{The plot title.}

\item{...}{Further parameters to geom_point()}
}
\value{
A ggplot object.
}
\description{
Exploratory c-QTc plot
}
\examples{
library(dplyr)
library(magrittr)

verapamil_cqtc \%>\%
  cqtc_plot(color = "ACTIVE")

dofetilide_cqtc \%>\%
filter(ACTIVE == 1) \%>\%
  cqtc_plot()

}
