% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_functions.R
\name{cqtc_prediction_dataset}
\alias{cqtc_prediction_dataset}
\title{Prediction data set}
\usage{
cqtc_prediction_dataset(fit, cqtc, level = 0.9, method = "boot")
}
\arguments{
\item{fit}{A model fit object.}

\item{cqtc}{The qtcf object.}

\item{level}{The confidence level.}

\item{method}{The CI calculation method ("profile" or "boot").}
}
\value{
A data frame.
}
\description{
`r lifecycle::badge("experimental")`

This approach follows Parkinson et al.
}
