% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_exploration.R
\name{cqtc_time_course_plot}
\alias{cqtc_time_course_plot}
\title{Time course plot for QT parameter and drug concentration}
\usage{
cqtc_time_course_plot(obj, param = "QTCF", title = "")
}
\arguments{
\item{obj}{A cqtc object.}

\item{param}{The parameter to be plotted, as character.}

\item{title}{The plot title.}
}
\value{
A ggplot object.
}
\description{
Time course plot for QT parameter and drug concentration
}
