% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_functions.R
\name{derive_group_delta}
\alias{derive_group_delta}
\title{Add delta to reference group by NTIME}
\usage{
derive_group_delta(cqtc, parameter = "DQTCF", reference_filter = "ACTIVE == 0")
}
\arguments{
\item{cqtc}{A cqtc data set.}

\item{parameter}{The parameter to calculate the delta over. Defaults to DQTCF.}

\item{reference_filter}{The filter term to identify the control group.}
}
\value{
A cqtc object.
}
\description{
Add delta to reference group by NTIME
}
