% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_functions.R
\name{derive_hr}
\alias{derive_hr}
\title{Derive HR from RR}
\usage{
derive_hr(obj, silent = NULL)
}
\arguments{
\item{obj}{A cqtc object.}

\item{silent}{Suppress messages.}
}
\value{
A cqtc object with the HR field added.
}
\description{
Derive HR from RR
}
\examples{
derive_hr(dofetilide_cqtc)
dplyr::select(dofetilide_cqtc, -"HR") |>
  derive_hr()
}
