% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_functions.R
\name{derive_rr}
\alias{derive_rr}
\title{Derive RR from HR}
\usage{
derive_rr(obj, silent = NULL)
}
\arguments{
\item{obj}{A cqtc object.}

\item{silent}{Suppress messages.}
}
\value{
A cqtc object with the RR field added.
}
\description{
Derive RR from HR
}
\examples{
derive_rr(dofetilide_cqtc)
dplyr::select(dofetilide_cqtc, -"RR") |>
  derive_rr()
}
