% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cqtc.R
\name{eg_nif_from_sdtm}
\alias{eg_nif_from_sdtm}
\title{Make EG nif object from SDTM data}
\usage{
eg_nif_from_sdtm(sdtm, extrt, eg_ntime = NULL, pctestcd = NULL, silent = NULL)
}
\arguments{
\item{sdtm}{A sdtm object.}

\item{silent}{Suppress messages, as logical.}
}
\value{
A data frame.
}
\description{
Make EG nif object from SDTM data
}
