% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_exploration.R
\name{find_inconsistent_rr_hr}
\alias{find_inconsistent_rr_hr}
\title{Identify entries with inconsistent RR and HR fields.}
\usage{
find_inconsistent_rr_hr(obj, threshold = 0.1)
}
\arguments{
\item{obj}{A cqtc object.}

\item{threshold}{The allowed fractional deviation.}
}
\value{
A data frame with entries in which the reported RR and the RR
re-calculated from the reported HR (i.e., rr_recalc) are different by the
specified threshold.
}
\description{
Identify entries with inconsistent RR and HR fields.
}
\examples{
find_inconsistent_rr_hr(dofetilide_cqtc, 0.01)
}
