% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cqtc.R
\name{find_time_matched}
\alias{find_time_matched}
\title{Find matching ECG and concentration measurements}
\usage{
find_time_matched(
  nif,
  conc_analyte = NULL,
  ecg_analyte = "QTCF",
  time_window = 10/60
)
}
\arguments{
\item{nif}{A nif object with observations for the conc_analyte and the
ecg_analyte.}

\item{conc_analyte}{The concentration analyte.}

\item{ecg_analyte}{The ECG parameter analyte.}

\item{time_window}{The allowed time window.}
}
\description{
Finds time-matched pairs between ECG (QTCF) measurements and concentration
(PK) measurements based on proximity in time.
}
