% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_class.R
\name{hash}
\alias{hash}
\title{Generic hash function}
\usage{
hash(obj)
}
\arguments{
\item{obj}{A cqtc object.}
}
\value{
The XXH128 hash of the nif object as character.
}
\description{
Generic hash function
}
