% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_class.R
\name{hash_cqtc}
\alias{hash_cqtc}
\title{Generate the XXH128 hash of a cqtc object}
\usage{
hash_cqtc(obj)
}
\arguments{
\item{obj}{A catc object.}
}
\value{
The XXH128 hash of the catc object as character.
}
\description{
Generate the XXH128 hash of a cqtc object
}
