% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_exploration.R
\name{hr_by_time_plot}
\alias{hr_by_time_plot}
\title{Plot HR over time}
\usage{
hr_by_time_plot(
  obj,
  param = "HR",
  group = NULL,
  title = "",
  size = 0.5,
  alpha = 1,
  lwd = 0.6
)
}
\arguments{
\item{obj}{A cqtc object.}

\item{param}{The parameter to plot.}

\item{group}{The grouping variable.}

\item{title}{The plot title.}

\item{size}{Point size.}

\item{alpha}{The alpha value for points.}

\item{lwd}{The line width.}
}
\value{
A ggplot2 object.
}
\description{
Plot HR over time
}
