% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_exploration.R
\name{hr_plot}
\alias{hr_plot}
\title{Plot ECG parameter by HR}
\usage{
hr_plot(obj, param = "QTCF", fit = TRUE, method = "lm")
}
\arguments{
\item{obj}{A cqtc object.}

\item{param}{The parameter to plot.}

\item{fit}{Show regression fit.}

\item{method}{The method for geom_smooth.}
}
\value{
A ggplot object.
}
\description{
Plot ECG parameter by HR
}
