% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_exploration.R
\name{hysteresis_plot}
\alias{hysteresis_plot}
\title{Title}
\usage{
hysteresis_plot(obj, param = "QTCF", ...)
}
\arguments{
\item{obj}{A cqtc object.}

\item{param}{The parameter to plot on the y axis.}

\item{...}{Further parameters to geom_point.}
}
\value{
A ggplot object.
}
\description{
Title
}
