% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_functions.R
\name{is_hr_rr_consistent}
\alias{is_hr_rr_consistent}
\title{Test whether the HR and RR fields are consistent with each other}
\usage{
is_hr_rr_consistent(obj, threshold = 0.05, silent = NULL)
}
\arguments{
\item{obj}{A cqtc object.}

\item{threshold}{The allowed relative difference between the original RR and
the RR as recalculated from HR, defaults to 0.05.}

\item{silent}{Suppress messages.}
}
\value{
A logical value.
}
\description{
Test whether the HR and RR fields are consistent with each other
}
\examples{
is_hr_rr_consistent(dofetilide_cqtc)
}
