% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_class.R
\name{plot.cqtc}
\alias{plot.cqtc}
\title{Print function for cqtc object}
\usage{
\method{plot}{cqtc}(x, ...)
}
\arguments{
\item{x}{A cqtc object.}

\item{...}{Further parameters.}
}
\value{
A ggplot object.
}
\description{
Print function for cqtc object
}
