% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quinidine_cqtc}
\alias{quinidine_cqtc}
\title{Quinidine c-QTc data set}
\format{
A cqtc object, i.e., a class wrapper around a data frame with 1008 rows and
38 columns:

\describe{
  \item{ID}{Subject ID}
  \item{NTIME}{Nominal time in hours}
  \item{Dofetilide}{Dofetilide plasma concentration in ng/ml}
  \item{QT}{The QT interval in ms}
  \item{QTCF}{The QTcF interval in ms}
  \item{DQTCF}{The change from baseline in QTCF in ms}
  \item{RR}{The RR interval in ms}
  \item{HR}{Heart rate, derived from RR in the original data set}
  \item{BL_QTCF}{Baseline QTCF, derived from QTCF at NTIME -0.5}
  ...
}
}
\source{
Derived from <https://physionet.org/content/ecgrdvq/1.0.0/#files-panel>
}
\usage{
quinidine_cqtc
}
\description{
A concentration-QTc analysis set for quinidine with triplicate observations,
derived from 'SCR_002_Clinical_Data'.
}
\references{
Johannesen L, Vicente J, Mason JW, Sanabria C, Waite-Labott K,
 Hong M, Guo P, Lin J, Sørensen JS, Galeotti L, Florian J, Ugander M,
 Stockbridge N, Strauss DG. Differentiating Drug-Induced Multichannel Block
 on the Electrocardiogram: Randomized Study of Dofetilide, Quinidine,
 Ranolazine, and Verapamil. Clin Pharmacol Ther. 2014 Jul 23. doi:
 10.1038/clpt.2014.155.
}
\keyword{datasets}
