% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_class.R
\name{subjects.cqtc}
\alias{subjects.cqtc}
\title{Subjects in a cqtc object}
\usage{
\method{subjects}{cqtc}(obj)
}
\arguments{
\item{obj}{A cqtc object.}
}
\value{
A data frame.
}
\description{
Subjects in a cqtc object
}
\examples{
subjects(dofetilide_cqtc)

}
