% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_col_param}
\alias{validate_col_param}
\title{Validate character parameter}
\usage{
validate_col_param(
  value,
  df = NULL,
  allow_null = FALSE,
  allow_multiple = FALSE
)
}
\arguments{
\item{value}{The parameter value.}

\item{df}{A data frame the column names of which must contain the value(s),
if not NULL.}

\item{allow_null}{Allow NULL values, as logical.}

\item{allow_multiple}{Allow character vectors, as logical.}
}
\value{
Nothing.
}
\description{
Validate character parameter
}
