% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{verapamil}
\alias{verapamil}
\title{Verapamil c-QTc data set}
\format{
A data frame with 704 rows and 7
columns:
\describe{
  \item{ID}{Subject ID}
  \item{ACTIVE}{Active treatment}
  \item{NTIME}{Nominal time in hours}
  \item{Cplasma}{Dofetilide plasma concentration in ng/ml}
  \item{QT}{The QT interval in ms}
  \item{QTCF}{The QTcF interval in ms}
  \item{RR}{The RR interval in ms}
}
}
\source{
<https://github.com/joannaparkinson/C-QTc-hands-on-tutorial>
}
\usage{
verapamil
}
\description{
A concentration-QTc analysis set for dofetilide.
}
\references{
Parkinson, J., Dota, C. & Rekić, D. Practical guide to
concentration-QTc modeling: a hands-on tutorial. J Pharmacokinet Pharmacodyn
52, 43 (2025). https://doi.org/10.1007/s10928-025-09981-8
}
\keyword{datasets}
