% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{verapamil_cqtc}
\alias{verapamil_cqtc}
\title{Verapamil c-QTc data set}
\format{
A cqtc object, i.e., a class wrapper around a data frame with 704 rows and 7
columns:
\describe{
  \item{ID}{Subject ID}
  \item{NTIME}{Nominal time in hours}
  \item{Verapamil}{Verapamil plasma concentration in ng/ml}
  \item{QT}{The QT interval in ms}
  \item{QTCF}{The QTcF interval in ms}
  \item{DQTCF}{The change from baseline in QTCF in ms}
  \item{RR}{The RR interval in ms}
  \item{HR}{Heart rate, derived from RR in the original data set}
  ...
}
}
\source{
<https://github.com/joannaparkinson/C-QTc-hands-on-tutorial>
}
\usage{
verapamil_cqtc
}
\description{
A concentration-QTc analysis set for verapamil
}
\references{
Parkinson, J., Dota, C. & Rekić, D. Practical guide to
concentration-QTc modeling: a hands-on tutorial. J Pharmacokinet Pharmacodyn
52, 43 (2025). https://doi.org/10.1007/s10928-025-09981-8
}
\keyword{datasets}
