% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_functions.R
\name{add_bl_popmean}
\alias{add_bl_popmean}
\title{Add population mean column}
\usage{
add_bl_popmean(cqtc, param = "QTCF")
}
\arguments{
\item{cqtc}{A cqtc object.}

\item{param}{The column name for the baseline value.}
}
\value{
a cqtc object with the population mean column(n) added for the
specified parameters. The name(s) follow the naming "PM_xx", where "xx" is
the parameter name.
}
\description{
Add population mean column
}
