% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_make.R
\name{auto_cqtc}
\alias{auto_cqtc}
\title{Title}
\usage{
auto_cqtc(
  sdtm,
  treatment = NULL,
  pc_analyte = NULL,
  eg_analyte = c(QTCF = "QTCF", HR = "HR", RR = "RR", QT = "QTGDUR"),
  egcat = "CENTRAL ECG",
  ntime_method = "TPT",
  NTIME_lookup = NULL,
  silent = NULL,
  verbose = TRUE
)
}
\arguments{
\item{sdtm}{A sdtm object.}

\item{treatment}{The EXTRT.}

\item{pc_analyte}{The PCTESTCD.}

\item{eg_analyte}{The EGTESTCD.}

\item{egcat}{The EGCAT to filter observations for.}

\item{ntime_method}{The field to derive the nominal time from. Allowed values
are 'TPT', 'TPTNUM' and 'ELTM'. Defaults to xxTPT where xx is the domain
name.}

\item{NTIME_lookup}{A data frame with two columns, a column that defines the
custom nominal time information in the target domain (e.g., 'PCELTM'), and
'NTIME'. This data frame is left_joined into the observation data frame
to provide the NTIME field.}

\item{silent}{Suppress messages.}

\item{verbose}{Verbose messages.}
}
\value{
A cqtc object.
}
\description{
Title
}
