% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_class.R
\name{cqtc}
\alias{cqtc}
\title{Create cqtc object}
\usage{
cqtc(obj = NULL, conc_field = NULL, silent = NULL, rr_threshold = 0.1)
}
\arguments{
\item{obj}{A data frame.}

\item{conc_field}{The field in the input data set that represents the
independent concentration variable, if 'CONC' is not provided.}

\item{silent}{Suppress warnings, as logical.}

\item{rr_threshold}{The allowed fractional deviation between the recorded RR
interval and the RR interval back-calculated from the recorded HR.}
}
\value{
A cqtc object.
}
\description{
Create cqtc object
}
