% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_exploration.R
\name{cqtc_model_plot}
\alias{cqtc_model_plot}
\title{Plot predicted dQTcF over concentration}
\usage{
cqtc_model_plot(
  obj,
  mod,
  x_label = "concentration (ng/ml)",
  y_label = "dQTcF",
  title = NULL,
  level = 0.9,
  size = 2,
  alpha = 0.1,
  lwd = 0.6,
  loess = FALSE,
  refline = NULL
)
}
\arguments{
\item{obj}{A cqtc object.}

\item{mod}{A linear model.}

\item{x_label}{The x axis label.}

\item{y_label}{The y axis label.}

\item{title}{The plot title.}

\item{level}{The prediction interval.}

\item{size}{Point size.}

\item{alpha}{Alpha for points.}

\item{lwd}{Line width for point range.}

\item{loess}{Show LOESS, as logical.}

\item{refline}{Plot horizontal dashed reference lines at thse y axis values,
dafaults to NULL (no lines).}
}
\value{
A ggplot2 object.
}
\description{
Plot predicted dQTcF over concentration
}
