% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_exploration.R
\name{cqtc_model_table}
\alias{cqtc_model_table}
\title{Tabulate model parameters}
\usage{
cqtc_model_table(mod)
}
\arguments{
\item{mod}{The linear mixed-effects model}
}
\value{
A data frame.
}
\description{
Tabulate model parameters
}
