% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_exploration.R
\name{cqtc_plot}
\alias{cqtc_plot}
\title{Exploratory c-QTc plot}
\usage{
cqtc_plot(obj, param = "QTCF", color = NULL, fit = TRUE, method = "lm", ...)
}
\arguments{
\item{obj}{A cqtc object.}

\item{param}{The parameter to plot on the y axis.}

\item{color}{The column to be used for coloring.}

\item{fit}{Show regression fit.}

\item{method}{The method for geom_smooth.}

\item{...}{Further parameters to geom_point()}
}
\value{
A ggplot object.
}
\description{
Exploratory c-QTc plot
}
\examples{
library(dplyr)
library(magrittr)

verapamil_cqtc \%>\%
  cqtc_plot(color = "ACTIVE")

dofetilide_cqtc \%>\%
filter(ACTIVE == 1) \%>\%
  cqtc_plot()

}
