% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_exploration.R
\name{cqtc_plot}
\alias{cqtc_plot}
\title{Title}
\usage{
cqtc_plot(obj, param = "QTCF", fit = TRUE, method = "lm", ...)
}
\arguments{
\item{obj}{A cqtc object.}

\item{param}{The parameter to plot on the y axis.}

\item{fit}{Show regression fit.}

\item{method}{The method for geom_smooth.}

\item{...}{Further parameters to geom_point()}
}
\value{
A ggplot object.
}
\description{
Title
}
