% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqtc_class.R
\name{new_cqtc}
\alias{new_cqtc}
\title{Create cqtc object}
\usage{
new_cqtc(obj = NULL, conc_field = NULL, silent = NULL, rr_inconsistency = 0.1)
}
\arguments{
\item{obj}{A data frame.}
}
\value{
A cqtc object.
}
\description{
`r lifecycle::badge("deprecated")`
}
